Minor changes to the data processing
